+++ packages/xai/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @zhivex-ai/xai
 
+## 0.2.7-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @zhivex-ai/core@1.23.0-next.0
+  - @zhivex-ai/openai@0.13.5-next.0
+
 ## 0.2.6
 
 ### Patch Changes